requirements.txt
streamlit
pandas
numpy
torch
sentence-transformers
faiss-cpu
scikit-learn
pymupdf
textract
python-docx

